+++ Changes.md
@@ -33,6 +33,7 @@ Improvements
   - Increased menu item icon sizes.
   - A lock icon is now displayed next to read-only nodes.
 - RenderPassEditor : Changed the current render pass indicator to yellow to match other context-related UI elements.
+- GraphEditor : Moved "Show Input Connections" and "Show Output Connections" to "Connections" sub-menu and added "Show Input Labels" and "Show Output Labels" items.
 
 Fixes
 -----